Sorted out the charge animation issue. Seems to be working fine now

+++ DSP2803x_examples_ccsv5/i2c_eeprom/User/User_Prototypes.h
@@ -54,7 +54,7 @@ unsigned char CRC8(unsigned char *ptr, unsigned char len,unsigned char key);
 
 void Battery_Status(void);
 
-float Charging_Animation(float real_SOC);
+int Charging_Animation(int real_SOC);
 
 //#pragma CODE_SECTION(DELAY_US, "ramfuncs");
